started adding background color for items depending on quality

+++ app/src/main/java/com/BlizzardArmory/diablo/Items/Shoulders.java
@@ -24,13 +24,13 @@ public class Shoulders {
     private String tooltipParams;
     @SerializedName("requiredLevel")
     @Expose
-    private Integer requiredLevel;
+    private Double requiredLevel;
     @SerializedName("itemLevel")
     @Expose
-    private Integer itemLevel;
+    private Double itemLevel;
     @SerializedName("stackSizeMax")
     @Expose
-    private Integer stackSizeMax;
+    private Double stackSizeMax;
     @SerializedName("accountBound")
     @Expose
     private Boolean accountBound;
@@ -48,13 +48,13 @@ public class Shoulders {
     private Double armor;
     @SerializedName("attacksPerSecond")
     @Expose
-    private Integer attacksPerSecond;
+    private Double attacksPerSecond;
     @SerializedName("minDamage")
     @Expose
-    private Integer minDamage;
+    private Double minDamage;
     @SerializedName("maxDamage")
     @Expose
-    private Integer maxDamage;
+    private Double maxDamage;
     @SerializedName("slots")
     @Expose
     private String slots;
@@ -66,13 +66,13 @@ public class Shoulders {
     private AttributesHtml attributesHtml;
     @SerializedName("openSockets")
     @Expose
-    private Integer openSockets;
+    private Double openSockets;
     @SerializedName("set")
     @Expose
     private Set set;
     @SerializedName("seasonRequiredToDrop")
     @Expose
-    private Integer seasonRequiredToDrop;
+    private Double seasonRequiredToDrop;
     @SerializedName("dye")
     @Expose
     private Dye dye;
@@ -123,27 +123,27 @@ public class Shoulders {
         this.tooltipParams = tooltipParams;
     }
 
-    public Integer getRequiredLevel() {
+    public Double getRequiredLevel() {
         return requiredLevel;
     }
 
-    public void setRequiredLevel(Integer requiredLevel) {
+    public void setRequiredLevel(Double requiredLevel) {
         this.requiredLevel = requiredLevel;
     }
 
-    public Integer getItemLevel() {
+    public Double getItemLevel() {
         return itemLevel;
     }
 
-    public void setItemLevel(Integer itemLevel) {
+    public void setItemLevel(Double itemLevel) {
         this.itemLevel = itemLevel;
     }
 
-    public Integer getStackSizeMax() {
+    public Double getStackSizeMax() {
         return stackSizeMax;
     }
 
-    public void setStackSizeMax(Integer stackSizeMax) {
+    public void setStackSizeMax(Double stackSizeMax) {
         this.stackSizeMax = stackSizeMax;
     }
 
@@ -187,27 +187,27 @@ public class Shoulders {
         this.armor = armor;
     }
 
-    public Integer getAttacksPerSecond() {
+    public Double getAttacksPerSecond() {
         return attacksPerSecond;
     }
 
-    public void setAttacksPerSecond(Integer attacksPerSecond) {
+    public void setAttacksPerSecond(Double attacksPerSecond) {
         this.attacksPerSecond = attacksPerSecond;
     }
 
-    public Integer getMinDamage() {
+    public Double getMinDamage() {
         return minDamage;
     }
 
-    public void setMinDamage(Integer minDamage) {
+    public void setMinDamage(Double minDamage) {
         this.minDamage = minDamage;
     }
 
-    public Integer getMaxDamage() {
+    public Double getMaxDamage() {
         return maxDamage;
     }
 
-    public void setMaxDamage(Integer maxDamage) {
+    public void setMaxDamage(Double maxDamage) {
         this.maxDamage = maxDamage;
     }
 
@@ -235,11 +235,11 @@ public class Shoulders {
         this.attributesHtml = attributesHtml;
     }
 
-    public Integer getOpenSockets() {
+    public Double getOpenSockets() {
         return openSockets;
     }
 
-    public void setOpenSockets(Integer openSockets) {
+    public void setOpenSockets(Double openSockets) {
         this.openSockets = openSockets;
     }
 
@@ -251,11 +251,11 @@ public class Shoulders {
         this.set = set;
     }
 
-    public Integer getSeasonRequiredToDrop() {
+    public Double getSeasonRequiredToDrop() {
         return seasonRequiredToDrop;
     }
 
-    public void setSeasonRequiredToDrop(Integer seasonRequiredToDrop) {
+    public void setSeasonRequiredToDrop(Double seasonRequiredToDrop) {
         this.seasonRequiredToDrop = seasonRequiredToDrop;
     }
